FunDay/requirements.txt
pandas==1.4.3
nltk==3.7
numpy==1.23.5
tqdm==4.64.0
scikit-learn==1.1.2
gensim==4.2.0
spacy==3.4.3
streamlit==1.23.0
transformers==4.29.2
umap-learn==0.5.3
wget==3.2
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.9.3
torchvision==0.13.1
pymupdf==1.22.3
docker==6.1.2
pipwin==0.5.2
plotly==5.10.0
elasticsearch==8.6.2
streamlit-extras==0.2.7
streamlit-plotly-events==0.0.6
wordcloud==1.9.2

